Fixes background color bug scaling bug
Our default behvaiour when exporting is to set the background color of the button in Interface Builder. This is fine when only considering a single file. However, when reusing an instance of a button in IB with various ToggleButton subclasses, the background color cannot abide solely by the IB variable.

This fix changes the behaviour so that it makes the background dynamically sync with the defaults of the current custom view.

It also disables interaction on the custom view itself.